pyproject.toml
[project]
name = "fuhrerschein-app"
version = "0.1.0"
description = "German driving theory practice app with AI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "psycopg2-binary>=2.9.9",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.4",
]

[tool.hatch.build.targets.wheel]
packages = ["."]